[ESOBOO6 - iOS]  Fix feedback from the code review
