+++ ios-plugin/swift-lang/src/main/java/io/ecocode/ios/swift/checks/idleness/IdleTimerDisabledCheck.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package io.ecocode.ios.swift.checks.environment.idleness;
+package io.ecocode.ios.swift.checks.idleness;
 
 import io.ecocode.ios.swift.RegisterRule;
 import io.ecocode.ios.swift.Swift;
