+++ ios-plugin/swift-lang/src/test/java/io/ecocode/ios/swift/checks/sobriety/TorchFreeCheckTests.java
