+++ ios-plugin/swift-lang/src/test/resources/checks/sobriety/torchUsage_typeInference_on_trigger.swift
@@ -10,7 +10,6 @@ final class AppDelegate: NSObject, UIApplicationDelegate {
 
         let device = AVCaptureDevice.default(for: AVMediaType.video)
         device.torchMode = AVCaptureTorchMode.on
-   
         return true
     }
 }
